requirements.txt
requests
python-dotenv==0.19.2